Run process controller tests on Windows
